pyproject.toml
[project]
name = "agent-skill-linter"
version = "0.1.0"
description = "Lint agent skills for spec compliance and publishing readiness"
readme = "README.md"
authors = [
    { name = "William Yeh", email = "william.pjyeh@gmail.com" }
]
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "skills-ref",
    "click",
    "pyyaml",
    "rich",
]

[project.scripts]
skill-lint = "agent_skill_linter.cli:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 99
